feat : fix arch unit failures
