--- src/main/java/com/learning/mfscreener/models/portfolio/Transaction.java
@@ -1,6 +1,0 @@
-package com.learning.mfscreener.models.portfolio;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-
-public record Transaction(LocalDate txnDate, BigDecimal units, BigDecimal nav, BigDecimal tax) {}
